requirements.txt
flask>=3.1
python-dotenv
schedule
langextract
